+++ packages/frame3/CHANGELOG.md
@@ -1,5 +1,14 @@
 # frame3
 
+## 1.1.10
+
+### Patch Changes
+
+- return 不加.value && 新增 v()手动加.value
+- Updated dependencies
+  - frame3-compiler-sfc@1.1.6
+  - frame3-runtime-core@1.1.9
+
 ## 1.1.9
 
 ### Patch Changes